Move the chunk-aligned padding helper from the coaddition module to _array_utils as pad_dask_array_to_grid since it is a generic dask array utility with no mosaicking dependencies
